Move logic to the same module and remove seta.cfg
