--- tools/seta.cfg
@@ -1,7 +1,0 @@
-{
-    "admins": ["vaibhavmagarwal@gmail.com", "jmaher@mozilla.com", "sabergeass@gmail.com"],
-    "results": ["joel.maher@gmail.com"],
-    "hostname": "smtp.gmail.com:587",
-    "username": "setamozilla@gmail.com",
-    "password": ""
-}
